% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batch_process}
\alias{batch_process}
\title{Batch process database queries to improve performance}
\usage{
batch_process(ids, batch_size = 100, process_fn)
}
\arguments{
\item{ids}{Vector of IDs to process}

\item{batch_size}{Number of IDs to process in each batch}

\item{process_fn}{Function to process each batch}
}
\value{
Combined results from all batches
}
\description{
Batch process database queries to improve performance
}
