% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_all_stations_means_function.R
\name{calculate_all_stations_means}
\alias{calculate_all_stations_means}
\title{Calculate all stations mean of specified parameter}
\usage{
calculate_all_stations_means(
  data,
  parameter,
  threshold,
  total_days,
  verbose = FALSE
)
}
\arguments{
\item{station_name}{The name of the station to filter the data.}
}
\description{
Calculate all stations mean of specified parameter
}
