% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AOT40.R
\name{calculate_AOT40}
\alias{calculate_AOT40}
\title{Calculate the AOT40}
\usage{
calculate_AOT40(data, parameter, months)
}
\arguments{
\item{data}{A grouped data frame containing the data (in tbl format)}

\item{parameter}{The parameter to calculate the AOT40 for}

\item{months}{The months to calculate the AOT40 for}
}
\description{
Calculate the AOT40
}
