% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aot40_exceeding_stations.R
\name{calculate_aot40_exceeding_stations}
\alias{calculate_aot40_exceeding_stations}
\title{This function calculates the 8-hour rolling averages for all stations and identifies
the number of stations with daily maximum values exceeding 120 µg/m³ for a specified parameter.}
\usage{
calculate_aot40_exceeding_stations(parameter_name)
}
\arguments{
\item{parameter_name}{The name of the parameter (e.g., "O3", "PM10") to use for the calculations.}
}
\value{
A data frame with a single column \code{num_exceeding_stations} representing the number of stations with daily maximum 8-hour rolling averages exceeding 120 µg/m³.
}
\description{
This function calculates the 8-hour rolling averages for all stations and identifies
the number of stations with daily maximum values exceeding 120 µg/m³ for a specified parameter.
}
