% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AQI.R
\name{calculate_AQI}
\alias{calculate_AQI}
\title{calculate AQI}
\usage{
calculate_AQI(data, parameter)
}
\arguments{
\item{data}{A grouped data frame containing the data (in tbl format)}

\item{parameter}{The parameter to be analyzed.}
}
\description{
calculate AQI
}
