% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average.R
\name{calculate_average}
\alias{calculate_average}
\title{Calculate the average of available data}
\usage{
calculate_average(data, parameter, data_in_year)
}
\arguments{
\item{data}{A grouped data frame containing the data (in tbl format)}

\item{parameter}{The parameter to calculate the average for}
}
\description{
Calculate the average of available data
}
