% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_co_exceeding_stations.R
\name{calculate_co_exceeding_stations}
\alias{calculate_co_exceeding_stations}
\title{Calculate CO Exceeding Stations for 8-hour Moving Average}
\usage{
calculate_co_exceeding_stations()
}
\value{
A data.frame with the stations and the annual average exceeding the threshold.
}
\description{
This function calculates the number of stations where the maximum daily 8-hour
moving average of CO exceeds 10 mg/m3. For valid calculation, at least 18
values per day should be present. If a station has less than 90\% valid data
over the year, its annual average is set to NA.
}
