% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_consecutive_exceedance.R
\name{calculate_consecutive_exceedance}
\alias{calculate_consecutive_exceedance}
\title{Calculate the consecutive exceedance of available data}
\usage{
calculate_consecutive_exceedance(
  data,
  parameter,
  threshold,
  consecutive,
  exceedance_direction
)
}
\arguments{
\item{data}{A grouped data frame containing the data (in tbl format)}

\item{parameter}{The parameter to calculate the exceedance for}

\item{threshold}{The threshold value to compare the data with}

\item{consecutive}{The number of consecutive exceedances}

\item{exceedance_direction}{The direction of the exceedance (above or below)}
}
\description{
Calculate the consecutive exceedance of available data
}
