% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_exceedance.R
\name{calculate_exceedance}
\alias{calculate_exceedance}
\title{Calculate the exceedance of available data}
\usage{
calculate_exceedance(data, parameter, threshold, exceedance_direction)
}
\arguments{
\item{data}{A grouped data frame containing the data (in tbl format)}

\item{parameter}{The parameter to calculate the exceedance for}

\item{threshold}{The threshold value to compare the data with}
}
\description{
Calculate the exceedance of available data
}
