% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_exceedance_days_daily.R
\name{calculate_exceedance_days_daily}
\alias{calculate_exceedance_days_daily}
\title{Calculate Exceedance Days for specified parameter from daily_detail}
\usage{
calculate_exceedance_days_daily(data, parameter, threshold)
}
\arguments{
\item{parameter}{The parameter for which the exceedance days are calculated.}

\item{threshold}{The threshold value for the parameter.}

\item{daily_data}{The daily detail data containing parameter values for each day.}
}
\value{
The number of days exceeding the specified threshold for the parameter.
}
\description{
This function calculates the number of days exceeding a specified threshold for a given parameter from daily_detail data.
}
