% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overall_average_by_city_threshold.R
\name{calculate_overall_average_by_city_threshold}
\alias{calculate_overall_average_by_city_threshold}
\title{Calculate Overall Average for specified parameter from daily_detail for all stations in a city}
\usage{
calculate_overall_average_by_city_threshold(
  parameter,
  data_type = "daily",
  threshold = 0,
  until_year = 2023
)
}
\arguments{
\item{parameter}{The parameter for which the overall average is calculated.}

\item{city_name}{The name of the city.}
}
\value{
The overall average value of the parameter for all stations in the city.
}
\description{
This function calculates the overall average value of a specified parameter from daily_detail data for all stations in a city.
}
