% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_average_PM25_by_city.R
\name{calculate_overall_PM25_average_by_city}
\alias{calculate_overall_PM25_average_by_city}
\title{Calculate Overall Average PM2.5 for specified parameter from daily_detail for all stations in a city}
\usage{
calculate_overall_PM25_average_by_city(
  data_type = "daily",
  threshold = 75,
  by_city = FALSE,
  until_year = 2023
)
}
\arguments{
\item{city_name}{The name of the city.}

\item{parameter}{The parameter for which the overall average is calculated.}
}
\value{
The overall average value of the parameter for all stations in the city.
}
\description{
This function calculates the overall average value of a specified parameter from daily_detail data for all stations in a city.
}
