% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Parameter_Mean_Function.R
\name{calculate_parameter_mean}
\alias{calculate_parameter_mean}
\title{Calculate mean of specified parameter}
\usage{
calculate_parameter_mean(
  data,
  parameter,
  threshold,
  total_days,
  verbose = FALSE
)
}
\arguments{
\item{station_name}{The name of the station to filter the data.}
}
\description{
Calculate mean of specified parameter
}
