% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentage.R
\name{calculate_percentage}
\alias{calculate_percentage}
\title{Calculate the percentage of available data}
\usage{
calculate_percentage(data, parameter, data_in_year, season = NULL)
}
\arguments{
\item{data}{A grouped data frame containing the data (in tbl format)}

\item{parameter}{The parameter to calculate the percentage for}

\item{data_in_year}{The number of data entries in a year}

\item{season}{Optional month numbers used for seasonal completeness.}
}
\description{
Calculate the percentage of available data
}
