% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_o3_rolling_averages.R
\name{calculate_rolling_8hour_avg}
\alias{calculate_rolling_8hour_avg}
\title{Calculate Rolling 8-Hour Averages for Ozone (O3)}
\usage{
calculate_rolling_8hour_avg()
}
\description{
This function retrieves O3 data from a given SQLite database,
and calculates rolling 8-hour averages for stations with 90\% or higher
data availability.
}
