% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rolling_8hour_avg_CO.R
\name{calculate_rolling_8hour_avg_CO}
\alias{calculate_rolling_8hour_avg_CO}
\title{Calculate 8-Hour Rolling Averages for CO}
\usage{
calculate_rolling_8hour_avg_CO()
}
\value{
A \code{data.frame} containing the station name (\code{Istasyon}), timestamp (\code{Tarih}), and
the calculated 8-hour rolling average for CO (\code{CO_8hour_avg}).
}
\description{
This function calculates the 8-hour rolling averages for CO levels in hourly air quality data.
The function only considers stations where data availability is 90\% or higher.
}
