% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_function.R
\name{check_page_errors}
\alias{check_page_errors}
\title{Check for page errors}
\usage{
check_page_errors(remDr, action = "general")
}
\arguments{
\item{remDr}{The remote driver object for the Selenium web browser.}

\item{action}{The action being performed (e.g., "download").}
}
\value{
A list containing the error status and message.
}
\description{
Check for page errors
}
\examples{
check_page_errors(remDr, "download")

}
