% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_function.R
\name{check_station_exists}
\alias{check_station_exists}
\title{Check if station exists}
\usage{
check_station_exists(remDr, bolge, sehir, istasyon)
}
\arguments{
\item{remDr}{The remote driver object for the Selenium web browser.}

\item{bolge}{The region to select.}

\item{sehir}{The city to select.}

\item{istasyon}{The station to select.}

\item{startdate}{The start date for the data download (format: "DD.MM.YYYY").}

\item{enddate}{The end date for the data download (format: "DD.MM.YYYY").}
}
\value{
TRUE if the station exists, FALSE otherwise.
}
\description{
Check if station exists
}
