% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_hourly_list_stations.R
\name{consecutive_hourly_list_stations}
\alias{consecutive_hourly_list_stations}
\title{List of stations consecutively exceeding a specified parameter from hourly_detail}
\usage{
consecutive_hourly_list_stations(parameter_name, threshold = 350)
}
\arguments{
\item{parameter_name}{The name of the desired parameter. For example, "PM10".}

\item{threshold}{The threshold value. Default value is 350 µg/m^3.}
}
\value{
List of stations consecutively exceeding the specified parameter for three hours and the total duration of exceedance.
}
\description{
This function finds stations that consecutively exceed a specified threshold value for a given parameter (e.g., PM10 level) for three consecutive hours.
It retrieves data from the hourly details table in the database to return a list of stations and the total duration of exceeding the threshold.
}
