% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_data.R
\name{convert_to_24h}
\alias{convert_to_24h}
\title{Convert 12-hour time to 24-hour format}
\usage{
convert_to_24h(hour, meridiem)
}
\arguments{
\item{hour}{Numeric hour (1-12)}

\item{meridiem}{"AM" or "PM"}
}
\value{
Numeric hour in 24-hour format (0-23)
}
\description{
Convert 12-hour time to 24-hour format
}
