% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_Stations_Function.R
\name{count_station_list}
\alias{count_station_list}
\title{Count Station List}
\usage{
count_station_list(processed_data, param_name)
}
\arguments{
\item{processed_data}{Processed data containing station information.}

\item{param_name}{Name of the parameter to check.}
}
\value{
Number of stations for the specified parameter.
}
\description{
Count Station List
}
