% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_stations.R
\name{count_stations}
\alias{count_stations}
\title{Count stations}
\usage{
count_stations(data, parameter, threshold)
}
\arguments{
\item{data}{The data with necessary columns (in tbl format)}

\item{parameter}{The parameter to calculate the average for}

\item{threshold}{The threshold for the percentage}
}
\description{
Count stations
}
