% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_stations_with_parameter_threshold.R
\name{count_stations_with_parameter_threshold}
\alias{count_stations_with_parameter_threshold}
\title{Count Stations meeting threshold criteria for specified parameter}
\usage{
count_stations_with_parameter_threshold(
  parameter_name,
  data_type = "daily",
  threshold = 90,
  verbose = FALSE,
  season = NULL
)
}
\arguments{
\item{parameter_name}{The name of the parameter}

\item{data_type}{Either 'daily' or 'hourly'}

\item{threshold}{The threshold percentage for data availability (default is 90)}
}
\description{
Count Stations meeting threshold criteria for specified parameter
}
