% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_excel.R
\name{create_analysis_excel}
\alias{create_analysis_excel}
\title{Create an Excel file for analysis}
\usage{
create_analysis_excel(data, description, additional = list())
}
\arguments{
\item{data}{The data to be written to Excel}

\item{description}{The description of the data}

\item{additional}{Additional views and their descriptions}
}
\description{
Create an Excel file for analysis
}
