% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dygraphs_function.R
\name{create_daily_time_series_graph}
\alias{create_daily_time_series_graph}
\title{Create Daily Dygraphs}
\usage{
create_daily_time_series_graph(data, station_name, parameters)
}
\arguments{
\item{data}{Dataframe containing data to be graphed}

\item{station_name}{specified station name}

\item{parameters}{parameters to be graphed}
}
\description{
Create Daily Dygraphs
}
