% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_with_multiple_stations.R
\name{create_hourly_time_series_graph_multiple_stations}
\alias{create_hourly_time_series_graph_multiple_stations}
\title{Create Dygraphs for Hourly Time Series}
\usage{
create_hourly_time_series_graph_multiple_stations(
  data,
  station_names,
  parameters
)
}
\arguments{
\item{parameters}{A character vector specifying the name(s) of the parameter(s) to be plotted.}

\item{query_result}{A dataframe containing data from a database query. It should include columns for "Tarih" (date/time) and the specified parameter(s) along with the station names.}

\item{station_name}{A character vector specifying the name of the station(s) for which the time series graph will be created.}
}
\value{
A Dygraph object displaying the hourly time series graph.
}
\description{
This function generates a Dygraph for hourly time series data, showing the variation of a specified parameter over time for a given station or multiple stations.
}
