% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_combinations.R
\name{create_parameter_view}
\alias{create_parameter_view}
\title{Create SQL view from parameter combination data}
\usage{
create_parameter_view(conn, view_name, data)
}
\arguments{
\item{conn}{PostgreSQL connection}

\item{view_name}{Name for the view}

\item{data}{Data to store in the view}
}
\value{
TRUE if view was created successfully
}
\description{
Create SQL view from parameter combination data
}
\keyword{internal}
