% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_conns.R
\name{create_postgres_conn}
\alias{create_postgres_conn}
\title{Create a PostgreSQL database connection}
\usage{
create_postgres_conn()
}
\value{
A PostgreSQL connection object
}
\description{
Create a PostgreSQL database connection
}
