% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite_base.R
\name{create_sqlite_base}
\alias{create_sqlite_base}
\title{Create SQLite schema for temiz-hava database}
\usage{
create_sqlite_base(db_path = "./temiz-hava.sqlite")
}
\arguments{
\item{db_path}{Path to the SQLite database file}
}
\description{
This function creates the necessary tables in the SQLite database for storing hourly and daily detail data.
}
\examples{
create_sqlite_base()
create_sqlite_base("./temiz-hava.sqlite")
}
