% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dygraphs_function.R
\name{create_station_graph}
\alias{create_station_graph}
\title{Create Dygraphs}
\usage{
create_station_graph(query_result, x_column, ...)
}
\arguments{
\item{query_result}{Dataframe containing data from database query.}

\item{station_name}{specified station name.}
}
\description{
Create Dygraphs
}
