% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_above_exceedance_days_double_threshold.R
\name{daily_above_exceedance_days_double_threshold}
\alias{daily_above_exceedance_days_double_threshold}
\title{Calculate Exceedance Days for specified parameter from daily_detail for all stations}
\usage{
daily_above_exceedance_days_double_threshold(
  parameter,
  threshold,
  exceedance_limit
)
}
\arguments{
\item{parameter}{The parameter for which the exceedance days are calculated.}

\item{threshold}{The threshold value for the parameter.}

\item{daily_data}{The daily detail data containing parameter values for each day.}
}
\value{
A data frame containing the number of days exceeding the specified threshold for the parameter for each station.
}
\description{
This function calculates the number of days exceeding a specified threshold for a given parameter from daily_detail data for all stations.
}
