% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_detail_load_from_database.R
\name{daily_detail_load_from_database}
\alias{daily_detail_load_from_database}
\title{Hourly detail load from database by station name}
\usage{
daily_detail_load_from_database(processed_data, station_name)
}
\arguments{
\item{station_name}{The name of the station to filter the data.}
}
\description{
Hourly detail load from database by station name
}
