% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daily_detail_savesqlite.R
\name{daily_detail_save_to_database}
\alias{daily_detail_save_to_database}
\title{Sum detail save to database}
\usage{
daily_detail_save_to_database(processed_data)
}
\arguments{
\item{processed_data}{processed data.}
}
\description{
Sum detail save to database
}
