% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_daily_stations.R
\name{daily_list_stations_with_parameter_count}
\alias{daily_list_stations_with_parameter_count}
\title{Count Stations for specified parameter from daily_detail}
\usage{
daily_list_stations_with_parameter_count(parameter_name)
}
\arguments{
\item{data}{A pair of numbers.}
}
\description{
Count Stations for specified parameter from daily_detail
}
