% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail_savesqlite.R
\name{detail_save_to_database}
\alias{detail_save_to_database}
\title{save detail data to database}
\usage{
detail_save_to_database(processed_data, table_name, verbose = FALSE)
}
\arguments{
\item{processed_data}{processed data.}
}
\description{
save detail data to database
}
