% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_conns.R
\name{disconnect_postgres}
\alias{disconnect_postgres}
\title{Disconnect from PostgreSQL database}
\usage{
disconnect_postgres(conn)
}
\arguments{
\item{conn}{The database connection object}
}
\value{
Boolean indicating success
}
\description{
Disconnect from PostgreSQL database
}
