% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_PM25.R
\name{estimate_PM25}
\alias{estimate_PM25}
\title{Estimate PM2.5}
\usage{
estimate_PM25(data, data_in_year)
}
\arguments{
\item{data}{A grouped data frame containing the data (in tbl format)}

\item{data_in_year}{The number of data entries in a year}
}
\description{
Estimate PM2.5
}
