% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceedance_days_3_threshold.R
\name{exceedance_days_3_threshold}
\alias{exceedance_days_3_threshold}
\title{Calculate exceedance days for a specified parameter}
\usage{
exceedance_days_3_threshold(
  parameter_name,
  threshold = 125,
  exceedance_count = 3
)
}
\arguments{
\item{parameter_name}{A character string specifying the parameter to check (e.g., "SO2", "NO2", etc.).}

\item{threshold}{A numeric value representing the threshold value (default is 125 µg/m³).}

\item{exceedance_count}{An integer representing the number of times the threshold should be exceeded (default is 3).}
}
\description{
This function calculates the number of days a specified parameter exceeds a given threshold value
more than a specified number of times. The parameter can be any valid pollutant column in the dataset.
}
