% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_function.R
\name{findElementWithRetry}
\alias{findElementWithRetry}
\title{Find element with retry}
\usage{
findElementWithRetry(driver, using, value, max_retries = 5)
}
\arguments{
\item{driver}{The remote driver object for the Selenium web browser.}

\item{using}{The method to use for finding the element (e.g., "id", "xpath").}

\item{value}{The value to use for finding the element.}

\item{max_retries}{The maximum number of retries to attempt.}
}
\value{
The found element.
}
\description{
Find element with retry
}
\examples{
findElementWithRetry(remDr, "id", "dropdown12-contentDataDowloadNew")

}
