% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_combinations.R
\name{generate_view_name}
\alias{generate_view_name}
\title{Generate view name for a parameter combination}
\usage{
generate_view_name(parameters, data_type, threshold)
}
\arguments{
\item{parameters}{Vector of parameter names}

\item{data_type}{Either 'daily' or 'hourly'}

\item{threshold}{Threshold percentage (not used in view name)}
}
\value{
A valid view name
}
\description{
Generate view name for a parameter combination
}
\keyword{internal}
