% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_stations_function.R
\name{get_all_station_hourly_detail_data}
\alias{get_all_station_hourly_detail_data}
\title{Get all data from the hourly_detail table for all stations}
\usage{
get_all_station_hourly_detail_data(db_conn)
}
\arguments{
\item{db_conn}{The SQLite database connection.}
}
\value{
A data frame containing all data from the hourly_detail table for all stations.
}
\description{
Get all data from the hourly_detail table for all stations
}
