% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_function.R
\name{get_dropdown_options}
\alias{get_dropdown_options}
\title{Get all available options from a dropdown}
\usage{
get_dropdown_options(remDr, css_selector)
}
\arguments{
\item{remDr}{The remote driver object for the Selenium web browser.}

\item{css_selector}{The CSS selector for the dropdown element.}
}
\value{
A character vector of all available options.
}
\description{
Get all available options from a dropdown
}
