% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_conns.R
\name{get_postgres_tables}
\alias{get_postgres_tables}
\title{Get the list of tables in the PostgreSQL database}
\usage{
get_postgres_tables(conn)
}
\arguments{
\item{conn}{The database connection object}
}
\value{
A character vector of table names
}
\description{
Get the list of tables in the PostgreSQL database
}
\examples{
get_postgres_tables(conn)
}
