% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkish_encodings.R
\name{get_problematic_cities}
\alias{get_problematic_cities}
\title{Get list of cities that need special handling}
\usage{
get_problematic_cities()
}
\value{
Character vector of city names that need special handling
}
\description{
Get list of cities that need special handling
}
