% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Parameter_Mean_40_Function.R
\name{get_station_name_mean_above_40}
\alias{get_station_name_mean_above_40}
\title{Calculate specified parameter of specified station with threshold}
\usage{
get_station_name_mean_above_40(
  all_query_result,
  parameter_name,
  threshold = 0.9
)
}
\arguments{
\item{station_name}{The name of the station to filter the data.}
}
\description{
Calculate specified parameter of specified station with threshold
}
