% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_conns.R
\name{get_table_row_count}
\alias{get_table_row_count}
\title{Check how many rows are in a table}
\usage{
get_table_row_count(conn, table_name)
}
\arguments{
\item{conn}{The database connection object}

\item{table_name}{The name of the table}
}
\value{
The number of rows in the table
}
\description{
Check how many rows are in a table
}
\examples{
get_table_row_count(conn, "daily_detail")
get_table_row_count(conn, "hourly_detail")
}
