% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_detail_load_from_database.R
\name{hourly_detail_load_from_database}
\alias{hourly_detail_load_from_database}
\title{Hourly detail load from database by station name}
\usage{
hourly_detail_load_from_database(station_name)
}
\arguments{
\item{station_name}{The name of the station to filter the data.}
}
\description{
Hourly detail load from database by station name
}
