% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hourly_detail_savesqlite.R
\name{hourly_detail_save_to_database}
\alias{hourly_detail_save_to_database}
\title{Hourly detail save to database}
\usage{
hourly_detail_save_to_database(processed_data, station_name)
}
\arguments{
\item{processed_data}{processed data.}
}
\description{
Hourly detail save to database
}
