% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stations_for_hourly_detail.R
\name{hourly_list_stations_with_parameter}
\alias{hourly_list_stations_with_parameter}
\title{List Stations for specified parameter from hourly_detail}
\usage{
hourly_list_stations_with_parameter(parameter_name)
}
\arguments{
\item{data}{A pair of numbers.}
}
\description{
List Stations for specified parameter from hourly_detail
}
