% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_stations_for_hourly_detail.R
\name{hourly_list_stations_with_parameter_count}
\alias{hourly_list_stations_with_parameter_count}
\title{Count Stations for specified parameter from hourly_detail}
\usage{
hourly_list_stations_with_parameter_count(parameter_name)
}
\arguments{
\item{data}{A pair of numbers.}
}
\description{
Count Stations for specified parameter from hourly_detail
}
