% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hourly_Station_Average_Function.R
\name{hourly_station_average}
\alias{hourly_station_average}
\title{Station Average for hourly data}
\usage{
hourly_station_average(parameter_name, threshold = 90)
}
\arguments{
\item{data}{station based air parameters data.}
}
\description{
Station Average for hourly data
}
