% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.temizhavaR.R
\name{init.temizhavaR}
\alias{init.temizhavaR}
\title{Init the settings for the temizhavaR package}
\usage{
init.temizhavaR(verbose = FALSE)
}
\arguments{
\item{verbose}{Print output if TRUE. Default is FALSE}
}
\description{
Init the settings for the temizhavaR package
}
