% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_conns.R
\name{is_postgres_connected}
\alias{is_postgres_connected}
\title{Check if database connection is valid}
\usage{
is_postgres_connected(conn)
}
\arguments{
\item{conn}{The database connection object}
}
\value{
Boolean indicating if connection is valid
}
\description{
Check if database connection is valid
}
