% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_AQI.R
\name{list_AQI}
\alias{list_AQI}
\title{list AQI by year for every station}
\usage{
list_AQI(parameter, data_type, threshold = 0, until_year = 2023)
}
\arguments{
\item{parameter}{The parameter to be analyzed.}

\item{data_type}{The type of data to be analyzed. Options are "daily" and "hourly".}

\item{until_year}{The year until which the data will be analyzed. Default is 2023.}

\item{theeshold}{The concentration threshold for the annual average. Stations with an annual average
below this value will be returned. (e.g., 40 µg/m³)}
}
\description{
list AQI by year for every station
}
