% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_count_station_name_treshold.R
\name{list_nb_stations_above_below_data_treshold}
\alias{list_nb_stations_above_below_data_treshold}
\title{List stations of specified parameter}
\usage{
list_nb_stations_above_below_data_treshold(parameter, threshold, mean)
}
\arguments{
\item{station_name}{The name of the station to filter the data.}
}
\description{
List stations of specified parameter
}
