% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_station_exceedance.R
\name{list_station_exceedance}
\alias{list_station_exceedance}
\title{List station exceedance}
\usage{
list_station_exceedance(
  data,
  parameter,
  threshold,
  data_threshold,
  threshold_direction,
  exceedance_threshold = 0
)
}
\arguments{
\item{data}{The data with necessary columns (in tbl format)}

\item{parameter}{The parameter to calculate the average for}

\item{threshold}{The threshold for the percentage}

\item{data_threshold}{The threshold for the percentage}

\item{threshold_direction}{The direction of the data threshold}

\item{exceedance_threshold}{The threshold for the exceedance}
}
\description{
List station exceedance
}
