% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stations_above_hourly_mean.R
\name{list_stations_above_hourly_mean}
\alias{list_stations_above_hourly_mean}
\title{Returns the list of stations with hourly average above a specified threshold for a given parameter.}
\usage{
list_stations_above_hourly_mean(parameter_name, threshold = 350)
}
\arguments{
\item{parameter_name}{The name of the parameter for stations.}

\item{threshold}{The threshold value for determining stations with yearly averages above it (default is 40).}
}
\value{
A dataframe containing the list of stations with yearly averages above the specified threshold for the given parameter.
}
\description{
Returns the list of stations with hourly average above a specified threshold for a given parameter.
}
