% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{memoize_with_timeout}
\alias{memoize_with_timeout}
\title{Create a time-limited memoized version of a function}
\usage{
memoize_with_timeout(fun, timeout = 3600)
}
\arguments{
\item{fun}{Function to memoize}

\item{timeout}{Time in seconds after which cache expires}
}
\value{
Memoized function with timeout
}
\description{
Create a time-limited memoized version of a function
}
