% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkish_encodings.R
\name{normalize_turkish}
\alias{normalize_turkish}
\title{Map Turkish characters to ASCII equivalents}
\usage{
normalize_turkish(text)
}
\arguments{
\item{text}{The text containing Turkish characters to be normalized.}
}
\value{
Normalized text with ASCII equivalents.
}
\description{
Map Turkish characters to ASCII equivalents
}
