% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_conns.R
\name{reconnect_postgres}
\alias{reconnect_postgres}
\title{Reconnect to PostgreSQL database if connection is lost}
\usage{
reconnect_postgres(conn)
}
\arguments{
\item{conn}{The database connection object}
}
\value{
A new connection object or NULL if reconnection fails
}
\description{
Reconnect to PostgreSQL database if connection is lost
}
