% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_function.R
\name{safeClick}
\alias{safeClick}
\title{Safe click on element}
\usage{
safeClick(driver, element)
}
\arguments{
\item{driver}{The remote driver object for the Selenium web browser.}

\item{element}{The element to click.}
}
\value{
None.
}
\description{
Safe click on element
}
\examples{
safeClick(remDr, element)
}
