% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_names_fix_function.R
\name{select_option}
\alias{select_option}
\title{Download air quality data from the specified website}
\usage{
select_option(dropdown_id, option_text)
}
\arguments{
\item{bolge}{The region to select.}

\item{sehir}{The city to select.}

\item{istasyon}{The station to select..}
}
\description{
Download air quality data from the specified website
}
