% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_function.R
\name{selectDropdownOption}
\alias{selectDropdownOption}
\title{Select dropdown option}
\usage{
selectDropdownOption(driver, dropdown_xpath, option_text, max_retries = 5)
}
\arguments{
\item{driver}{The remote driver object for the Selenium web browser.}

\item{dropdown_xpath}{The XPath of the dropdown element.}

\item{option_text}{The text of the option to select.}

\item{max_retries}{The maximum number of retries to attempt.}
}
\value{
TRUE if the option was successfully selected.
}
\description{
Select dropdown option
}
\examples{
selectDropdownOption(remDr, "dropdown1-contentDataDowloadNew", "Ankara")

}
