% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_average.R
\name{station_average}
\alias{station_average}
\title{Station Average}
\usage{
station_average(
  parameter_name,
  data_type = "daily",
  threshold = 90,
  season = NULL,
  until_year = 2023
)
}
\arguments{
\item{parameter_name}{The name of the parameter.}

\item{threshold}{The threshold percentage for data availability (default is 90).}

\item{season}{The season for which data availability is calculated (default is NULL). Use "summer" or "winter".}
}
\value{
A data frame with station names, their data availability percentages, and parameter averages, sorted by data availability.
}
\description{
Station Average
}
