% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_data.R
\name{summarize_data_quality}
\alias{summarize_data_quality}
\title{Summarize data quality for a station}
\usage{
summarize_data_quality(data)
}
\arguments{
\item{data}{Data frame returned by check_station_data}
}
\description{
Summarize data quality for a station
}
