% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_air_quality_data.R
\name{tidy_air_quality_data}
\alias{tidy_air_quality_data}
\title{Tidy air quality data}
\usage{
tidy_air_quality_data(
  tbl_db,
  table_name,
  check_hourly = FALSE,
  con = NULL,
  update_original = TRUE
)
}
\arguments{
\item{tbl_db}{A database table to clean.}

\item{table_name}{The name of the table to clean.}

\item{check_hourly}{Whether to check for hourly data.}

\item{con}{A database connection.}

\item{update_original}{Whether to update the original table.}
}
\description{
Tidy air quality data
}
\examples{
tidy_air_quality_data(tbl_db, "table_name")
}
