% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkish_encodings.R
\docType{data}
\name{turkish_city_map}
\alias{turkish_city_map}
\title{Map of problematic Turkish city names}
\format{
An object of class \code{list} of length 7.
}
\usage{
turkish_city_map
}
\description{
Maps Turkish city names that cause encoding issues to their ASCII equivalents
}
\keyword{datasets}
