% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_excel.R
\name{write_output_to_excel}
\alias{write_output_to_excel}
\title{Save analysis output to Excel file using openxslx}
\usage{
write_output_to_excel(output, output_filename)
}
\arguments{
\item{output}{The list objects containing analysis result DFs}

\item{output_filename}{Name of the Excel file to write data in}
}
\description{
Save analysis output to Excel file using openxslx
}
