% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_daily_detail_load.R
\name{all_daily_detail_load_from_database}
\alias{all_daily_detail_load_from_database}
\title{Hourly detail load from database}
\usage{
all_daily_detail_load_from_database(parm)
}
\arguments{
\item{parameter}{The name of the parameter to return. All parameters are returned if not provided.}
}
\description{
Hourly detail load from database
}
