% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_above_exceedance_days_all_stations.R
\name{calculate_above_exceedance_days_all_stations}
\alias{calculate_above_exceedance_days_all_stations}
\title{Calculate Exceedance Days for Specified Parameter from Daily Data}
\usage{
calculate_above_exceedance_days_all_stations(
  parameter,
  threshold = 125,
  exceedance_count = 3
)
}
\arguments{
\item{parameter}{The parameter for which the exceedance days are calculated.}

\item{threshold}{The threshold value for the parameter. Default is 125.}

\item{exceedance_count}{The minimum number of times the threshold should be exceeded to be included in the result. Default is 3.}
}
\value{
A data frame containing stations and the number of times they exceed the specified threshold for the parameter.
}
\description{
This function calculates the number of times stations exceed a specified threshold
for a given parameter from daily_detail data.
}
