% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_above_exceedance_days_all_stations.R,
%   R/list_stations_above_exceedance.R
\name{calculate_above_exceedance_days_all_stations}
\alias{calculate_above_exceedance_days_all_stations}
\title{Calculate Stations with Annual Average Above Threshold and Their Exceedance Days}
\usage{
calculate_above_exceedance_days_all_stations(parameter, threshold)
}
\arguments{
\item{parameter}{The parameter for which the exceedance days are calculated.}

\item{threshold}{The threshold value for the parameter.}

\item{pollutant_threshold}{The concentration threshold for the annual average. Stations with an annual average
above this value will be returned. (e.g., 40 µg/m³)}

\item{data_threshold}{The minimum data availability percentage required for a station to be included in the analysis. Default is 90.}

\item{daily_data}{The daily detail data containing parameter values for each day.}
}
\value{
A data frame containing the stations with an annual average above the specified threshold and the number of exceedance days.

A data frame containing the number of days exceeding the specified threshold for the parameter for each station.
}
\description{
This function calculates the annual average concentration of a specified air quality parameter
for stations that have a data availability percentage above 90\%. It returns
the stations whose annual average exceeds a specified threshold along with the number of days
the parameter exceeded that threshold.

This function calculates the number of days exceeding a specified threshold for a given parameter from daily_detail data for all stations.
}
\examples{
# Calculate stations with annual average PM10 above 40 µg/m³ and at least 90\% data availability
result <- calculate_above_exceedance_days_all_stations("PM10", 40)
print(result)
}
