% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stations_above_exceedance.R
\name{calculate_above_exceedance_days_all_stations}
\alias{calculate_above_exceedance_days_all_stations}
\title{Calculate Exceedance Days for specified parameter from daily_detail for all stations}
\arguments{
\item{daily_data}{The daily detail data containing parameter values for each day.}

\item{parameter}{The parameter for which the exceedance days are calculated.}

\item{threshold}{The threshold value for the parameter.}
}
\value{
A data frame containing the number of days exceeding the specified threshold for the parameter for each station.
}
\description{
This function calculates the number of days exceeding a specified threshold for a given parameter from daily_detail data for all stations.
}
