% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AOT40_percentage.R
\name{calculate_AOT40_percentage}
\alias{calculate_AOT40_percentage}
\title{Calculate the AOT40 percentage of available data}
\usage{
calculate_AOT40_percentage(data, parameter, months, days)
}
\arguments{
\item{data}{A grouped data frame containing the data (in tbl format)}

\item{parameter}{The parameter to calculate the percentage for}
}
\description{
Calculate the AOT40 percentage of available data
}
