% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_aot40_vegatation_and_forest_protection.R
\name{calculate_AOT40_vegetation_and_forest_protection}
\alias{calculate_AOT40_vegetation_and_forest_protection}
\title{Calculate the total AOT40 values for vegetation protection within a specified date range}
\usage{
calculate_AOT40_vegetation_and_forest_protection(
  parameter_name,
  start_month,
  end_month
)
}
\arguments{
\item{parameter_name}{The name of the parameter (e.g., "O3") to use for the calculations.}

\item{start_month}{The starting month (e.g., "05" for May) for the calculation.}

\item{end_month}{The ending month (e.g., "07" for July) for the calculation.}
}
\value{
A data frame with columns: Istasyon, total_AOT40, exceeds_6000, exceeds_18000.
}
\description{
This function calculates the total AOT40 values for vegetation protection based on the hourly ozone data.
It evaluates whether the AOT40 value exceeds 6000 (µg/m³)*hours and 18000 (µg/m³)*hours within a specified date range.
}
