% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_city_average.R
\name{calculate_city_average}
\alias{calculate_city_average}
\title{Calculate the city average of available data}
\usage{
calculate_city_average(
  data,
  parameter,
  cities,
  threshold,
  second_threshold = 0
)
}
\arguments{
\item{data}{The data with necessary columns (in tbl format)}

\item{parameter}{The parameter to calculate the average for}

\item{cities}{The cities data}

\item{threshold}{The threshold for the percentage}
}
\description{
Calculate the city average of available data
}
