% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_city_average.R
\name{calculate_city_average}
\alias{calculate_city_average}
\title{Calculate the city average of available data}
\usage{
calculate_city_average(averages, parameter, percentages, stations)
}
\arguments{
\item{averages}{A grouped data frame containing the averages (in tbl format)}

\item{parameter}{The parameter to calculate the average for}

\item{percentages}{A data frame containing the percentages (in tbl format)}

\item{stations}{A data frame containing the stations (in tbl format)}
}
\description{
Calculate the city average of available data
}
