% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hourly_aot40_exceeding_stations.R
\name{calculate_hourly_aot40_exceeding_stations}
\alias{calculate_hourly_aot40_exceeding_stations}
\title{Calculate Hourly Exceeding Stations}
\usage{
calculate_hourly_aot40_exceeding_stations(
  parameter_name,
  hour_window = 1,
  months = c(4, 5, 6, 7, 8, 9),
  threshold = 180
)
}
\arguments{
\item{parameter_name}{A string representing the name of the parameter (pollutant) to analyze.}

\item{hour_window}{An integer representing the number of hours for the rolling average window. Default is 1.}

\item{months}{A vector of integers representing the months to include in the analysis. Default is c(4, 5, 6, 7, 8, 9) for April to September.
Set to NULL for the entire year.}

\item{threshold}{A numeric value representing the threshold value for the parameter. Default is 180.}
}
\value{
A data.frame with the number of stations that exceed the threshold for the specified parameter.
}
\description{
This function calculates the number of stations that exceed a specified threshold
for a given parameter over a specified hourly window and within certain months,
or for the entire year if no months are specified.
}
