% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Parameter_Mean_Function.R
\name{calculate_parameter_mean}
\alias{calculate_parameter_mean}
\title{Calculate specified parameter of specified station}
\usage{
calculate_parameter_mean(query_result, parameter_name, threshold = 0.9)
}
\arguments{
\item{station_name}{The name of the station to filter the data.}
}
\description{
Calculate specified parameter of specified station
}
