% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rolling_exceedance.R
\name{calculate_rolling_exceedance}
\alias{calculate_rolling_exceedance}
\title{Calculate the rolling exceedance of available data}
\usage{
calculate_rolling_exceedance(
  data,
  parameter,
  threshold,
  rolling_hours,
  con,
  period = "Day",
  months = NULL
)
}
\arguments{
\item{data}{A grouped data frame containing the data (in tbl format)}

\item{parameter}{The parameter to calculate the exceedance for}

\item{threshold}{The threshold value to compare the data with}

\item{rolling_hours}{The number of hours to calculate the rolling exceedance for}

\item{period}{The period to calculate the rolling exceedance for (Day, Week, Month, Year)}
}
\description{
Calculate the rolling exceedance of available data
}
