% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_stations.R
\name{count_stations}
\alias{count_stations}
\title{Count Stations}
\usage{
count_stations(processed_data, param_name)
}
\arguments{
\item{processed_data}{processed data.}

\item{param_name}{specified paramater name.}
}
\description{
Count Stations
}
