% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dygraphs_function.R
\name{create_hourly_time_series_graph}
\alias{create_hourly_time_series_graph}
\title{Create Dygraphs}
\usage{
create_hourly_time_series_graph(data, station_name, parameters)
}
\arguments{
\item{station_name}{specified station name.}

\item{query_result}{Dataframe containing data from database query.}
}
\description{
Create Dygraphs
}
