% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_count_stations_with_parameter_threshold.R
\name{daily_count_stations_with_parameter_threshold}
\alias{daily_count_stations_with_parameter_threshold}
\title{Count Stations for specified parameter from daily_detail}
\usage{
daily_count_stations_with_parameter_threshold(parameter_name, threshold = 90)
}
\arguments{
\item{parameter_name}{The name of the parameter.}

\item{threshold}{The threshold percentage for data availability (default is 90).}
}
\description{
Count Stations for specified parameter from daily_detail
}
