% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stations_for_daily_detail.R
\name{daily_list_stations_with_parameter}
\alias{daily_list_stations_with_parameter}
\title{List Stations for specified parameter from daily_detail}
\usage{
daily_list_stations_with_parameter(parameter_name)
}
\arguments{
\item{data}{A pair of numbers.}
}
\description{
List Stations for specified parameter from daily_detail
}
