% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_list_stations_with_parameter_threshold.R
\name{daily_list_stations_with_parameter_threshold}
\alias{daily_list_stations_with_parameter_threshold}
\title{List Stations for specified parameter from daily_detail}
\usage{
daily_list_stations_with_parameter_threshold(
  parameter_name,
  threshold = 90,
  season = NULL
)
}
\arguments{
\item{parameter_name}{The name of the parameter.}

\item{threshold}{The threshold percentage for data availability (default is 90).}

\item{season}{The specific season to consider (e.g., "summer" for nisan, mayıs, haziran, temmuz, ağustos, eylül).}
}
\description{
List Stations for specified parameter from daily_detail
}
