% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daily_Station_Average_Function.R
\name{daily_station_average}
\alias{daily_station_average}
\title{Station Average for daily data}
\usage{
daily_station_average(parameter_name, threshold = 90)
}
\arguments{
\item{data}{station based air parameters data.}
}
\description{
Station Average for daily data
}
