% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_function.R
\name{download_data}
\alias{download_data}
\title{Download air quality data from the specified website}
\usage{
download_data(
  bolge,
  sehir,
  istasyon,
  data_type,
  startdate,
  enddate,
  result_dir
)
}
\arguments{
\item{bolge}{The region to select.}

\item{sehir}{The city to select.}

\item{istasyon}{The station to select.}

\item{data_type}{The type of data to download ("hourly" or "daily").}

\item{startdate}{The start date for the data download (format: "DD.MM.YYYY").}

\item{enddate}{The end date for the data download (format: "DD.MM.YYYY").}

\item{result_dir}{The directory where the downloaded data will be saved.}
}
\value{
None. Downloads the data and saves it to the specified directory.
}
\description{
Download air quality data from the specified website
}
