% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_list_aot40_threshold.R
\name{hourly_list_aot40_threshold}
\alias{hourly_list_aot40_threshold}
\title{Calculate AOT40 for a given period}
\usage{
hourly_list_aot40_threshold(
  start_months,
  end_months,
  threshold = 90,
  data_threshold = 0
)
}
\arguments{
\item{threshold}{The threshold percentage for data availability (default is 90).}

\item{start_dates}{The start dates of the period in "YYYY-MM-DD" format.}

\item{end_dates}{The end dates of the period in "YYYY-MM-DD" format.}
}
\value{
A data frame with columns: Istasyon_modified, aot40_measured, measured_hours, aot40_estimated, period.
}
\description{
This function calculates the AOT40 for a specified period and filters stations
with data availability greater than or equal to the threshold.
}
\examples{
calculate_aot40_hourly_threshold(c("2021-04-01", "2021-07-01"), c("2021-06-30", "2021-09-30"),threshold = 90)
}
