% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_station_averages.R
\name{list_station_averages}
\alias{list_station_averages}
\title{List station averages}
\usage{
list_station_averages(
  data,
  parameter,
  data_threshold,
  threshold_direction,
  threshold = 0
)
}
\arguments{
\item{data}{The data with necessary columns (in tbl format)}

\item{parameter}{The parameter to calculate the average for}

\item{data_threshold}{The threshold for the percentage}

\item{threshold_direction}{The direction of the data threshold}
}
\description{
List station averages
}
