% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stations_below_data_threshold.R
\name{list_stations_below_data_threshold}
\alias{list_stations_below_data_threshold}
\title{Returns the list of stations with yearly average below a specified threshold for a given parameter.}
\usage{
list_stations_below_data_threshold(parameter_name, threshold = 40)
}
\arguments{
\item{parameter_name}{The name of the parameter for stations.}

\item{threshold}{The threshold value for determining stations with yearly averages below it (default is 40).}
}
\value{
A dataframe containing the list of stations with yearly averages below the specified threshold for the given parameter.
}
\description{
Returns the list of stations with yearly average below a specified threshold for a given parameter.
}
