% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stations_with_parameter.R
\name{list_stations_with_parameter}
\alias{list_stations_with_parameter}
\title{List Stations for specified parameter from daily or hourly detail, grouped by year}
\usage{
list_stations_with_parameter(
  parameter_name,
  data_type = "daily",
  threshold = 0,
  data_threshold = 0,
  season = NULL,
  until_year = 2023
)
}
\arguments{
\item{parameter_name}{Name of the parameter to count stations for}

\item{data_type}{Either 'daily' or 'hourly'}

\item{threshold}{Minimum percentage of available data required (default: 0)}
}
\description{
List Stations for specified parameter from daily or hourly detail, grouped by year
}
