% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stations_with_parameter_count.R
\name{list_stations_with_parameter_count}
\alias{list_stations_with_parameter_count}
\title{Count Stations for specified parameter from daily or hourly detail, grouped by year}
\usage{
list_stations_with_parameter_count(parameter_name, data_type = "daily")
}
\arguments{
\item{parameter_name}{Name of the parameter to count stations for}

\item{data_type}{Either 'daily' or 'hourly'}
}
\description{
Count Stations for specified parameter from daily or hourly detail, grouped by year
}
