% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_Average_Function.R
\name{station_average}
\alias{station_average}
\title{Station Average for daily data}
\usage{
station_average(parameter_name, threshold = 90)
}
\arguments{
\item{data}{station based air parameters data.}
}
\description{
Station Average for daily data
}
