% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_Average_Function.R
\name{station_average}
\alias{station_average}
\title{Station Average}
\usage{
station_average(data, parameter_name)
}
\arguments{
\item{data}{station based air parameters data.}
}
\description{
Station Average
}
