% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_function.R
\name{wait_for_element}
\alias{wait_for_element}
\title{Wait for an element to appear on the page}
\usage{
wait_for_element(remDr, selector, type = "css", timeout = 10, interval = 0.5)
}
\arguments{
\item{remDr}{The remote driver object for the Selenium web browser.}

\item{selector}{The CSS selector or XPath to find the element.}

\item{type}{The type of selector ("css" or "xpath").}

\item{timeout}{Maximum time to wait in seconds.}

\item{interval}{Time between checks in seconds.}
}
\value{
TRUE if the element was found, FALSE otherwise.
}
\description{
Wait for an element to appear on the page
}
\examples{
wait_for_element(remDr, "#dropdown12-contentDataDowloadNew")

}
